Moved import linked file patterns preferences to FilePreferencesTab
