+++ src/main/java/org/jabref/gui/preferences/FileTab.fxml
@@ -10,6 +10,9 @@
 <?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
+<?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.RowConstraints?>
 <fx:root prefWidth="650.0" spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx/8.0.212" xmlns:fx="http://javafx.com/fxml/1"
          fx:controller="org.jabref.gui.preferences.FileTabView">
@@ -37,6 +40,12 @@
     </HBox>
     <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
 
+    <Label styleClass="sectionHeader" text="%Autosave"/>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
+        <Button fx:id="autosaveLocalLibrariesHelp"/>
+    </HBox>
+
     <Label styleClass="sectionHeader" text="%External file links"/>
     <HBox alignment="CENTER_LEFT" spacing="10.0">
         <Label text="%Main file directory"/>
@@ -62,9 +71,22 @@
     <CheckBox fx:id="searchFilesOnOpen" text="%When opening file link, search for matching file if no link is defined"/>
     <CheckBox fx:id="openBrowseOnCreate" text="%Automatically open browse dialog when creating new file link"/>
 
-    <Label styleClass="sectionHeader" text="%Autosave"/>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
-        <Button fx:id="autosaveLocalLibrariesHelp"/>
-    </HBox>
+    <Label styleClass="sectionHeader" text="%Linked file name conventions"/>
+    <GridPane hgap="4.0" vgap="4.0">
+        <columnConstraints>
+            <ColumnConstraints hgrow="SOMETIMES"/>
+            <ColumnConstraints hgrow="SOMETIMES"/>
+        </columnConstraints>
+        <rowConstraints>
+            <RowConstraints vgrow="SOMETIMES"/>
+            <RowConstraints vgrow="SOMETIMES"/>
+        </rowConstraints>
+        <Label text="%Filename format pattern"/>
+        <ComboBox fx:id="fileNamePattern" promptText="%Choose pattern" GridPane.columnIndex="1" editable="true"
+                  prefWidth="300" minWidth="300" maxWidth="300"/>
+
+        <Label text="%File directory pattern" GridPane.rowIndex="1"/>
+        <TextField fx:id="fileDirectoryPattern" GridPane.columnIndex="1" GridPane.rowIndex="1"
+                   prefWidth="300" minWidth="300" maxWidth="300"/>
+    </GridPane>
 </fx:root>
