+++ src/main/java/org/jabref/gui/preferences/ImportTab.fxml
@@ -1,33 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
 <fx:root prefWidth="650.0" spacing="10.0" type="VBox" xmlns="http://javafx.com/javafx/11.0.1"
          xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.ImportTabView">
 
     <Label styleClass="titleHeader" text="%Import"/>
 
-    <Label styleClass="sectionHeader" text="%Linked file name conventions"/>
-    <GridPane hgap="4.0" vgap="4.0">
-        <columnConstraints>
-            <ColumnConstraints hgrow="SOMETIMES"/>
-            <ColumnConstraints hgrow="SOMETIMES"/>
-        </columnConstraints>
-        <rowConstraints>
-            <RowConstraints vgrow="SOMETIMES"/>
-            <RowConstraints vgrow="SOMETIMES"/>
-        </rowConstraints>
-        <Label text="%Filename format pattern"/>
-        <ComboBox fx:id="fileNamePattern" promptText="%Choose pattern" GridPane.columnIndex="1" editable="true"
-                  prefWidth="300" minWidth="300" maxWidth="300"/>
-
-        <Label text="%File directory pattern" GridPane.rowIndex="1"/>
-        <TextField fx:id="fileDirectoryPattern" GridPane.columnIndex="1" GridPane.rowIndex="1"
-                   prefWidth="300" minWidth="300" maxWidth="300"/>
-    </GridPane>
 </fx:root>
