+++ src/main/java/org/jabref/gui/preferences/ImportTabView.java
@@ -1,18 +1,11 @@
 package org.jabref.gui.preferences;
 
-import javafx.fxml.FXML;
-import javafx.scene.control.ComboBox;
-import javafx.scene.control.TextField;
-
 import org.jabref.logic.l10n.Localization;
 import org.jabref.preferences.JabRefPreferences;
 
 import com.airhacks.afterburner.views.ViewLoader;
 
 public class ImportTabView extends AbstractPreferenceTabView<ImportTabViewModel> implements PreferencesTab {
-
-    @FXML private ComboBox<String> fileNamePattern;
-    @FXML private TextField fileDirectoryPattern;
 
     public ImportTabView(JabRefPreferences preferences) {
         this.preferences = preferences;
@@ -24,10 +17,6 @@ public class ImportTabView extends AbstractPreferenceTabView<ImportTabViewModel>
 
     public void initialize() {
         this.viewModel = new ImportTabViewModel(preferences);
-
-        fileNamePattern.valueProperty().bindBidirectional(viewModel.fileNamePatternProperty());
-        fileNamePattern.itemsProperty().bind(viewModel.defaultFileNamePatternsProperty());
-        fileDirectoryPattern.textProperty().bindBidirectional(viewModel.fileDirPatternProperty());
     }
 
     @Override
